+++ tusLibros/src/processor/MerchantProcessor.java
@@ -2,5 +2,5 @@ package processor;
 
 public interface MerchantProcessor {
 
-  void debit(double price, Integer number, String expirationDate, String owner);
+  void debit(double price, Long number, String expirationDate, String owner);
 }
